requirements.txt
argh==0.26.2
beautifulsoup4==4.12.2
browser_cookie3==0.16.2
html2text==2020.1.16
python-dotenv==1.0.0
requests==2.27.1
tabulate==0.9.0
termcolor==1.1.0
tomlkit==0.12.3
